+++ lib/screens/air_quality_detail_screen.dart
@@ -6,7 +6,7 @@ class AirQualityDetailScreen extends StatelessWidget {
 
   const AirQualityDetailScreen({super.key, required this.airQuality});
 
-  Color _getAqiColor(int aqi) {
+  Color _getAqiColor(double aqi) {
     if (aqi <= 50) return Colors.green;
     if (aqi <= 100) return Colors.yellow;
     if (aqi <= 150) return Colors.orange;
